refactoring Theme and Components to use more Helpers and remove unnecessary styles
