+++ App/Containers/Example/ExampleScreenStyle.js
@@ -1,53 +1,32 @@
 import { StyleSheet } from 'react-native'
-import Fonts from 'App/Theme/Fonts'
-import ApplicationStyles from 'App/Theme/ApplicationStyles'
+import { Helpers, Metrics, Fonts, Colors } from 'App/Theme'
 
 export default StyleSheet.create({
-  container: {
-    ...ApplicationStyles.screen.container,
-    margin: 30,
-    flex: 1,
-    justifyContent: 'center',
-  },
-  title: {
-    ...Fonts.style.h2,
+  error: {
+    ...Fonts.normal,
+    color: Colors.error,
+    marginBottom: Metrics.tiny,
     textAlign: 'center',
-    marginBottom: 10,
-  },
-  text: {
-    ...Fonts.style.normal,
-    textAlign: 'center',
-    marginBottom: 5,
   },
   instructions: {
-    ...Fonts.style.normal,
-    textAlign: 'center',
-    marginBottom: 5,
+    ...Fonts.normal,
     fontStyle: 'italic',
-  },
-  loading: {
-    ...Fonts.style.normal,
+    marginBottom: Metrics.tiny,
     textAlign: 'center',
-    marginBottom: 5,
-  },
-  result: {
-    ...Fonts.style.normal,
-    textAlign: 'center',
-    marginBottom: 5,
-  },
-  error: {
-    ...Fonts.style.normal,
-    textAlign: 'center',
-    marginBottom: 5,
-    color: 'red',
   },
   logoContainer: {
-    width: '100%',
+    ...Helpers.fullWidth,
     height: 300,
     marginBottom: 25,
   },
-  logo: {
-    width: '100%',
-    height: '100%',
+  result: {
+    ...Fonts.normal,
+    marginBottom: Metrics.tiny,
+    textAlign: 'center',
+  },
+  text: {
+    ...Fonts.normal,
+    marginBottom: Metrics.tiny,
+    textAlign: 'center',
   },
 })
