+++ App/Containers/Root/RootScreen.js
@@ -2,10 +2,10 @@ import React, { Component } from 'react'
 import NavigationService from 'App/Services/NavigationService'
 import AppNavigator from 'App/Navigators/AppNavigator'
 import { View } from 'react-native'
-import styles from './RootScreenStyle'
 import { connect } from 'react-redux'
 import StartupActions from 'App/Stores/Startup/Actions'
 import { PropTypes } from 'prop-types'
+import { Helpers } from 'App/Theme'
 
 class RootScreen extends Component {
   componentDidMount() {
@@ -15,7 +15,7 @@ class RootScreen extends Component {
 
   render() {
     return (
-      <View style={styles.container}>
+      <View style={Helpers.fill}>
         <AppNavigator
           // Initialize the NavigationService (see https://reactnavigation.org/docs/en/navigating-without-navigation-prop.html)
           ref={(navigatorRef) => {
