--- App/Containers/Root/RootScreenStyle.js
@@ -1,8 +1,0 @@
-import { StyleSheet } from 'react-native'
-import ApplicationStyles from 'App/Theme/ApplicationStyles'
-
-export default StyleSheet.create({
-  container: {
-    ...ApplicationStyles.screen.container,
-  },
-})
