+++ App/Containers/SplashScreen/SplashScreenStyle.js
@@ -1,21 +1,13 @@
 import { StyleSheet } from 'react-native'
-import Colors from 'App/Theme/Colors'
-import ApplicationStyles from 'App/Theme/ApplicationStyles'
+import { Colors } from 'App/Theme'
 
 export default StyleSheet.create({
   container: {
-    ...ApplicationStyles.screen.container,
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
     backgroundColor: Colors.primary,
   },
   logo: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
+    backgroundColor: Colors.white,
     height: 70,
     width: 70,
-    backgroundColor: 'white',
   },
 })
