--- App/Theme/ApplicationStyles.js
@@ -1,13 +1,0 @@
-/**
- * This file defines the base application styles.
- *
- * Use it to define generic component styles (e.g. the default text styles, default button styles...).
- */
-
-export default {
-  screen: {
-    container: {
-      flex: 1,
-    },
-  },
-}
