+++ App/Theme/Colors.js
@@ -7,7 +7,8 @@
 
 export default {
   transparent: 'rgba(0,0,0,0)',
-  //Example colors:
+  // Example colors:
+  white: '#ffffff',
   text: '#212529',
   primary: '#007bff',
   success: '#28a745',
