+++ App/Theme/index.js
@@ -2,7 +2,6 @@ import Colors from './Colors'
 import Fonts from './Fonts'
 import Metrics from './Metrics'
 import Images from './Images'
-import ApplicationStyles from './ApplicationStyles'
 import Helpers from './Helpers'
 
-export { Colors, Fonts, Images, Metrics, ApplicationStyles, Helpers }
+export { Colors, Fonts, Images, Metrics, Helpers }
